apps/ai-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
google-generativeai==0.8.4
python-multipart==0.0.20
pillow==11.1.0
pytesseract==0.3.13
scikit-learn==1.6.1
numpy==2.2.2
httpx==0.28.1
python-dotenv==1.0.1
pdfplumber==0.11.4
